Add output for initialized local variables.
ExprToken::output() has been partially implemented. Still to do:
function calls and operators.

Variables are flagged as not able to be put into a register if
their address is used via the address-of operator, like &var_name.
These variables are forced to go on the stack.

When we return from a function call, parameters that were before
the return address because they didn't fit in registers A through D
are popped off the stack to save the caller the trouble.